Allow the extension to be active in untrusted workspaces
Closes #305

Just ran into this today. At an absolute minimum, the language definition and grammar should be available even in untrusted workspaces, but a quick look suggests that the generate command is also safe. There is no substitution in the template, plus the user would have to run the command manually. At worst, untrusted files could set their own `.editorconfig`, which only applies to files below in the directory hierarchy: This is probably wanted behavior even in an untrusted workspace.